aider: Created a script to find available subnets within a specified range based on existing subnets and required size.
# Aider chat conversation:

USER: Create a script using python and netaddr in bin/available-subnets to help me figure out subnets I can use. I want to give it a range (e.g., 10.7.0.0/17), a list of existing subnets (e.g., 10.7.0.32/23, 10.7.43.54/28) and have it dump out a list of available subnets. So, between 10.7.0.32/23 and 10.7.43.54/28 I can fit up 1000 ips. how would I do that?

ASSISTANT: To create the script `available-subnets` in Python using the `netaddr` library, I will provide you with the diff for a new file. This script will take a base range and a list of existing subnets as input and output a list of available subnets that can fit a specified number of IPs.

```diff
--- /dev/null
+++ bin/available-subnets
@@ -0,0 +1,40 @@
+#!/usr/bin/env python3
+import sys
+from netaddr import IPNetwork, IPAddress
+
+def find_available_subnets(base_range, existing_subnets, required_size):
+    base_net = IPNetwork(base_range)
+    existing_nets = [IPNetwork(subnet) for subnet in existing_subnets]
+    available_ips = []
+
+    for ip in base_net.iter_hosts():
+        if not any(ip in subnet for subnet in existing_nets):
+            available_ips.append(ip)
+            if len(available_ips) == required_size:
+                yield IPNetwork('%s/%s' % (available_ips[0], base_net.prefixlen))
+                available_ips = []
+
+def main():
+    if len(sys.argv) < 4:
+        print("Usage: available-subnets <base_range> <required_size> <existing_subnet1> [<existing_subnet2> ...]")
+        sys.exit(1)
+
+    base_range = sys.argv[1]
+    required_size = int(sys.argv[2])
+    existing_subnets = sys.argv[3:]
+
+    available_subnets = find_available_subnets(base_range, existing_subnets, required_size)
+
+    for subnet in available_subnets:
+        print(subnet)
+
+if __name__ == '__main__':
+    main()
+
+# Make the script executable
+import os
+import stat
+filepath = os.path.join(os.getcwd(), 'bin/available-subnets')
+st = os.stat(filepath)
+os.chmod(filepath, st.st_mode | stat.S_IEXEC)
```

This script defines a function `find_available_subnets` that yields available subnets of the required size within the given base range, excluding the existing subnets. The `main` function handles command-line arguments and prints out the available subnets.

To use this script, you would run it from the command line like this:

```bash
./bin/available-subnets 10.7.0.0/17 1000 10.7.0.32/23 10.7.43.54/28
```

This will print out the available subnets within the 10.7.0.0/17 range that can fit 1000 IPs and are not overlapping with 10.7.0.32/23 and 10.7.43.54/28.

Please note that you need to have the `netaddr` library installed for this script to work. You can install it using `pip install netaddr`. Also, ensure that the script is executable by running `chmod +x bin/available-subnets`.