aider: Updated `open_subnets` to assign `new_range` to `track_open[position][1]` if its size is equal to `required_ips`.
# Aider chat conversation:

USER: If the new_range value is the size of required_ips then assign it to track_open[position] in the string (the second element).

ASSISTANT: Here is the diff to update the `open_subnets` function to assign the `new_range` to `track_open[position]` if the size of `new_range` is equal to `required_ips`:

```diff
--- bin/available-subnets
+++ bin/available-subnets
@@ ... @@
 def open_subnets(base_range, used_subnets, ranges):
   taken_subnets = list(used_subnets)
   # Take the order the user gives by default
   #ranges.sort(reverse=True)

-  track_open = {i + 1: [ranges[i], ''] for i in ranges(len(ranges))}
+  track_open = {i + 1: [ranges[i], ''] for i in range(len(ranges))}

   for position, (range_size, _) in track_open.items():
     required_ips = 2**(32 - range_size)
     new_range = find_range(base_range, taken_subnets, required_ips)
+    if len(new_range) == required_ips:
+      track_open[position][1] = str(IPNetwork(new_range[0]).cidr)

   # ... rest of the function ...
```

This diff corrects the `track_open` dictionary comprehension and adds the logic to assign the `new_range` to `track_open[position][1]` if the condition is met. Note that the `str(IPNetwork(new_range[0]).cidr)` is used to convert the first IP in the `new_range` list to its CIDR notation, which is the expected format for the subnet.