aider: Update `open_subnets` to include all IP addresses in `new_range` within a new `IPNetwork` object and append it to `taken_subnets`.
# Aider chat conversation:

USER: taken_subnets is an array of IPNetwork objects with each of the ip addresses as an IPAddress object. It looks like the new_subnet is only grabbing the first ip, not all of them.

ASSISTANT: Here is the diff to update the `open_subnets` function to ensure that `new_subnet` includes all the IP addresses in the `new_range` as `IPAddress` objects within an `IPNetwork` object:

```diff
--- bin/available-subnets
+++ bin/available-subnets
@@ ... @@
 def open_subnets(base_range, used_subnets, ranges):
   taken_subnets = list(used_subnets)
   track_open = {ix: [x, ''] for ix, x in enumerate(ranges)}

   for pos, range in enumerate(ranges):
     required_ips = 2**(32 - range)
     new_range = find_range(base_range, taken_subnets, required_ips)

     if len(new_range) != required_ips:
       continue

-    new_subnet = IPNetwork(new_range[0]).cidr
-    taken_subnets.append(new_subnet)
-    track_open[pos][1] = str(IPNetwork(new_range[0]).cidr)
+    new_subnet = IPNetwork(str(new_range[0]) + '/' + str(range))
+    for ip in new_range:
+        new_subnet.add_host(ip)
+    taken_subnets.append(new_subnet)
+    track_open[pos][1] = str(new_subnet.cidr)

   return [track_open[pos][1] for pos in track_open if track_open[pos][1]]
```

This diff updates the `open_subnets` function to create a new `IPNetwork` object for `new_subnet` with the correct CIDR notation, and then it adds all the IP addresses from `new_range` as hosts to this network. Finally, it appends the `new_subnet` to the `taken_subnets` list and updates the `track_open` dictionary with the CIDR notation of the new subnet.