aider: Update find_range to return an empty array if an IP exists in a taken_subnet and available_ips is less than required_ips.
# Aider chat conversation:

USER: In find_range, the first if in the loop needs to return an empty array if the ip exists in a taken_subnet and available_ips is less than required_ips.

ASSISTANT: Here is the diff to update the `find_range` function in the `bin/available-subnets` file:

```diff
--- bin/available-subnets
+++ bin/available-subnets
@@ ... @@
 def find_range(base_range, taken_subnets, required_ips):
   available_ips = []

   for ip in base_range.iter_hosts():
-    if any(ip in subnet for subnet in taken_subnets):
+    if any(ip in subnet for subnet in taken_subnets) and len(available_ips) < required_ips:
       return []

     if len(available_ips) == required_ips:
       return available_ips

     available_ips.append(ip)

   return available_ips
```

This change ensures that if an IP address is found within any of the `taken_subnets`, the function will return an empty list if the number of `available_ips` collected so far is less than the `required_ips`.